+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# wrapr 2.0.6 2021-01-29
+
+ * Add bc() blank-separated parsing, suggested by Emil Erik Pula Bellamy Begtrup-Bright \url{https://github.com/WinVector/wrapr/issues/12}.
+
 # wrapr 2.0.6 2020-12-06
 
  * Bump version.